requirements.txt
feedparser>=6.0
requests>=2.28
pyyaml>=6.0
